+++ subprocess/subprocess.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
+	"regexp"
 	"strconv"
 	"strings"
 
@@ -81,6 +82,24 @@ func Output(cmd *Cmd) (string, error) {
 	}
 
 	return strings.Trim(string(out), " \n"), err
+}
+
+var shellWordRe = regexp.MustCompile(`\A[A-Za-z0-9_@.-]+\z`)
+
+// ShellQuote returns a copied string slice where each element is quoted
+// suitably for sh.
+func ShellQuote(strs []string) []string {
+	dup := make([]string, 0, len(strs))
+
+	for _, str := range strs {
+		// Quote anything that looks slightly complicated.
+		if shellWordRe.FindStringIndex(str) == nil {
+			dup = append(dup, "'"+strings.Replace(str, "'", "'\\''", -1)+"'")
+		} else {
+			dup = append(dup, str)
+		}
+	}
+	return dup
 }
 
 func Trace(name string, args ...string) {
