+++ tools/str_tools.go
@@ -46,22 +46,6 @@ func QuotedFields(s string) []string {
 	return out
 }
 
-// ShellQuote returns a copied string slice where each element is quoted
-// suitably for sh.
-func ShellQuote(strs []string) []string {
-	dup := make([]string, 0, len(strs))
-
-	for _, str := range strs {
-		// Quote anything that looks slightly complicated.
-		if shellWordRe.FindStringIndex(str) == nil {
-			dup = append(dup, "'"+strings.Replace(str, "'", "'\\''", -1)+"'")
-		} else {
-			dup = append(dup, str)
-		}
-	}
-	return dup
-}
-
 // Ljust returns a copied string slice where each element is left justified to
 // match the width of the longest element in the set.
 func Ljust(strs []string) []string {
@@ -129,8 +113,7 @@ func Indent(str string) string {
 }
 
 var (
-	tabRe       = regexp.MustCompile(`(?m)^[ \t]+`)
-	shellWordRe = regexp.MustCompile(`\A[A-Za-z0-9_@.-]+\z`)
+	tabRe = regexp.MustCompile(`(?m)^[ \t]+`)
 )
 
 // Undent removes all leading tabs in the given string "str", line-wise.
